cleaned up unnecessary files, package, renamed files for clarity, reduced number of launch files by using arguments as was done in controls
