--- catkin_ws/src/auv_description/CMakeLists.txt
@@ -1,5 +1,0 @@
-cmake_minimum_required(VERSION 3.0.2)
-project(auv_description)
-
-find_package(catkin REQUIRED)
-catkin_package()
