--- catkin_ws/src/auv_description/launch/auv_description.launch
@@ -1,10 +1,0 @@
-<launch>
-    <arg name="model" default="$(find auv_description)/urdf/auv.urdf"/>
-    <arg name="gui" default="true"/>
-    <arg name="rvizconfig" default="$(find auv_description)/rviz/auv.rviz"/>
-
-    <param name="robot_description" command="cat $(arg model)"/>
-
-    <node name="rviz" pkg="rviz" type="rviz" args="-d $(arg rvizconfig)" required="true"/>
-    <node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher"/>
-</launch>
