--- catkin_ws/src/auv_description/package.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0"?>
-<package format="2">
-  <name>auv_description</name>
-  <version>0.1.0</version>
-  <description>urdf description and visualisation of auv</description>
-  <maintainer email="dev@mcgillrobotics.com">McGill Robotics</maintainer>
-  <license>GPLv3</license>
-  <url type="repository">https://github.com/mcgill-robotics/AUV-2020</url>
-
-  <buildtool_depend>catkin</buildtool_depend>
-
-  <exec_depend>robot_state_publisher</exec_depend>
-  <exec_depend>rviz</exec_depend>
-  <exec_depend>xacro</exec_depend>
-</package>
