--- catkin_ws/src/auv_description/urdf/auv.urdf
@@ -1,16 +1,0 @@
-<?xml version="1.0"?>
-<robot name="auv">
-    <link name="auv_base">
-        <visual>
-            <origin xyz="0 0 0" rpy="0 1.571 0"/>
-            <geometry>
-                <cylinder length="0.6" radius="0.2"/>
-            </geometry>
-        </visual>
-        <collision>
-            <geometry>
-                <cylinder length="0.6" radius="0.2"/>
-            </geometry>
-        </collision>
-    </link>
-</robot>
