+++ catkin_ws/src/bringup/launch/bringup.launch
@@ -1,9 +1,80 @@
 <launch>
+    <arg name="sim" default="false" />
+    <arg name="bag" default="false" />
+    <arg name="vision" default="true" />
 
-	<include file="$(find controls)/launch/controls.launch" />
-	<include file="$(find propulsion)/launch/propulsion.launch" />
-	<include file="$(find state_estimation)/launch/state_estimation.launch" />
+	<include file="$(find controls)/launch/controls.launch">
+		<arg name="sim" value="$(arg sim)" />
+	</include>
+	<include file="$(find propulsion)/launch/propulsion.launch">
+		<arg name="sim" value="$(arg sim)" />
+	</include>
+	<include file="$(find state_estimation)/launch/state_estimation.launch">
+		<arg name="sim" value="$(arg sim)" />
+	</include>
 
-	<include file="$(find vision)/launch/object_detection.launch" />
-	<!-- <include file="$(find planner)/launch/mission.launch" /> -->
+    <group if="$(arg vision)">
+		<include file="$(find vision)/launch/vision.launch">
+			<arg name="sim" value="$(arg sim)" />
+		</include>
+    </group>
+
+	<!-- this assumes planner is ran remotely (env set) after bringup, it will bag all the topics to have on the remote
+	laptop during robosub since it's inconvenient to get them off of the jetson -->
+    <group if="$(arg bag)">
+		<node name="bag_record_bringup" pkg="rosbag" type="record" output="screen" args=" 
+			/rosout
+			/rosout_agg
+
+			/tf
+			/tf_static
+
+			/error_quat
+			/state_quaternion_server/cancel
+			/state_quaternion_server/feedback
+			/state_quaternion_server/goal
+			/state_quaternion_server/result
+			/state_quaternion_server/status
+			/effort_server/cancel
+			/effort_server/feedback
+			/effort_server/goal
+			/effort_server/result
+			/effort_server/status
+			
+			/sbg/ekf_quat
+			/sbg/imu_data
+			/angular_velocity
+			/dead_reckon_report
+			/depth
+			/pose
+			/state_theta_x
+			/state_theta_y
+			/state_theta_z
+			/state_x
+			/state_y
+			/state_z
+
+			/global_x
+			/global_y
+			/global_z
+			/surge
+			/sway
+			/heave
+			/roll
+			/pitch
+			/yaw
+			/effort
+			/propulsion/thruster_forces
+			/propulsion/thruster_microseconds
+
+			/pid_quat_enable
+			/pid_x_enable
+			/pid_y_enable
+			/pid_z_enable
+			/x_setpoint
+			/y_setpoint
+			/z_setpoint
+			/quat_setpoint
+		"/>
+    </group>
 </launch>
