--- catkin_ws/src/controls/launch/pid-sim.launch
@@ -1,56 +1,0 @@
-<launch>
-	<node name="z_pid" pkg="pid" type="controller" >
-		<param name="Kp" value="20" /> 
-		<param name="Ki" value="2" />
-		<param name="Kd" value="30" />
-		<param name="upper_limit" value="50" />
-		<param name="lower_limit" value="-50" />
-		<param name="windup_limit" value="30" />
-		<param name="max_loop_frequency" value="100.0" />
-		<param name="min_loop_frequency" value="100.0" />
-
-		<param name="topic_from_controller" value="/global_z" />
-		<param name="topic_from_plant" value="/state_z" />
-		<param name="setpoint_topic" value="z_setpoint" />
-		<param name="pid_enable_topic" value="pid_z_enable" />
-	</node>
-
-	<node name="y_pid" pkg="pid" type="controller" >
-		<param name="Kp" value="20" /> 
-		<param name="Ki" value="0" />
-		<param name="Kd" value="10" />
-		<param name="upper_limit" value="10" />
-		<param name="lower_limit" value="-10" />
-		<param name="windup_limit" value="10" />
-		<param name="max_loop_frequency" value="100.0" />
-		<param name="min_loop_frequency" value="100.0" />
-
-		<param name="topic_from_controller" value="/global_y" />
-		<param name="topic_from_plant" value="/state_y" />
-		<param name="setpoint_topic" value="y_setpoint" />
-		<param name="pid_enable_topic" value="pid_y_enable" />
-	</node>
-
-	<node name="x_pid" pkg="pid" type="controller" >
-		<param name="Kp" value="20" /> 
-		<param name="Ki" value="0" />
-		<param name="Kd" value="10" />
-		<param name="upper_limit" value="10" />
-		<param name="lower_limit" value="-10" />
-		<param name="windup_limit" value="10" />
-		<param name="max_loop_frequency" value="100.0" />
-		<param name="min_loop_frequency" value="100.0" />
-
-		<param name="topic_from_controller" value="/global_x" />
-		<param name="topic_from_plant" value="/state_x" />
-		<param name="setpoint_topic" value="x_setpoint" />
-		<param name="pid_enable_topic" value="pid_x_enable" />
-	</node>
-
-	<node name="quaternion_pid" pkg="controls" type="quaternion_pid.py" respawn="True" output="screen">
-		<param name="Kp" value="0.2" />
-		<param name="Ki" value="0.0" />
-		<param name="Kd" value="0.2" />
-		<param name="windup_limit" value="0.5" />
-	</node>
-</launch>
