+++ catkin_ws/src/controls/launch/pid_config_gui.launch
@@ -1,3 +1,3 @@
 <launch>
-    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui" respawn="false" output="log" args = "--perspective-file $(find controls)/rqt-gui-config/pid_config.perspective"/>
+    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui" respawn="false" output="log" args = "--perspective-file $(find controls)/config/pid_config.perspective"/>
 </launch>
