--- catkin_ws/src/controls/launch/servers.launch
@@ -1,3 +1,0 @@
-<launch>
-	<node name="servers" pkg="controls" type="init_servers.py" respawn="false" output="screen"></node>
-</launch>
