+++ catkin_ws/src/controls/src/init_servers.py
@@ -2,14 +2,14 @@
 
 import rospy
 
-from servers.superimposer_server import SuperimposerServer
+from servers.effort_server import EffortServer
 from servers.state_quaternion_pid_server import StateQuaternionServer
 
 #define preempt callbacks using the cancel methods. This is necessary because action lib does not
 #allow methods to be callback function for preempting.
 if __name__ == "__main__":
     rospy.init_node("server manager")
-    sup = SuperimposerServer()
+    sup = EffortServer()
     sup.server.register_preempt_callback(lambda : sup.cancel())
     quat = StateQuaternionServer()
     quat.server.register_preempt_callback(lambda : quat.cancel())
