--- catkin_ws/src/planner/launch/bag_everything.launch
@@ -1,69 +1,0 @@
-<launch>
-<!-- 
-	this assumes planner is ran remotely (env set) after bringup, it will bag all the topics to have on the remote
-	laptop during robosub since it's inconvenient to get them off of the jetson
--->
-	<node name="bag_record_bringup" pkg="rosbag" type="record" output="screen" args=" 
-
-		/rosout
-		/rosout_agg
-
-		/effort
-		/propulsion/thruster_forces
-		/propulsion/thruster_microseconds
-
-		/angular_velocity
-		/dead_reckon_report
-		/depth
-		/pose
-		/pose_world
-		/reset_state_full
-		/reset_state_planar "/>
-		/sbg/ekf_quat
-		/sbg/imu_data
-		/state_theta_x
-		/state_theta_y
-		/state_theta_z
-		/state_x
-		/state_y
-		/state_z
-		/tf
-		/tf_static
-
-		/error_quat
-		/global_x
-		/global_y
-		/global_z
-		/heave
-		/pid_debug
-		/pid_quat_enable
-		/pid_x_enable
-		/pid_y_enable
-		/pid_z_enable
-		/pitch
-		/quat_setpoint
-		/roll
-		/state_quaternion_server/cancel
-		/state_quaternion_server/feedback
-		/state_quaternion_server/goal
-		/state_quaternion_server/result
-		/state_quaternion_server/status
-		/superimposer_server/cancel
-		/superimposer_server/feedback
-		/superimposer_server/goal
-		/superimposer_server/result
-		/superimposer_server/status
-		/surge
-		/sway
-		/x_pid/parameter_descriptions
-		/x_pid/parameter_updates
-		/x_setpoint
-		/y_pid/parameter_descriptions
-		/y_pid/parameter_updates
-		/y_setpoint
-		/yaw
-		/z_pid/parameter_descriptions
-		/z_pid/parameter_updates
-		/z_setpoint
-	"/>
-</launch>
