+++ catkin_ws/src/planner/launch/quali.launch
@@ -1,5 +1,6 @@
 <launch>
-	<include file="$(find bringup)/launch/bringup.launch" />
+	<include file="$(find bringup)/launch/bringup.launch">
+		<arg name="bag" value="true" />
+	</include>
 	<node pkg="planner" name="quali" type="quali.py" respawn="False" output="screen"/>
-	<include file="$(find planner)/launch/bag_everything.launch"/>
 </launch>
