+++ catkin_ws/src/propulsion/launch/propulsion.launch
@@ -1,4 +1,10 @@
 <launch>
-    <include file="$(find propulsion)/launch/thrust_mapper.launch" />
-    <include file="$(find propulsion)/launch/thrusters.launch" />
+    <arg name="sim" default="false" />
+    <group unless="$(arg sim)">
+        <node name="thrusters_serial_server" pkg="rosserial_python" type="serial_node.py" respawn="true">
+            <param name="port" value="/dev/auv_propulsion"/>
+            <param name="baud" value="57600"/>
+        </node>
+    </group>
+	<node name="thrust_mapper" pkg="propulsion" type="thrust_mapper.py" respawn="true" output="screen"/>
 </launch>
