--- catkin_ws/src/propulsion/launch/thrust_mapper.launch
@@ -1,3 +1,0 @@
-<launch>
-	<node name="thrust_mapper" pkg="propulsion" type="thrust_mapper.py" respawn="true" output="screen"/>
-</launch>
