--- catkin_ws/src/propulsion/launch/thrusters.launch
@@ -1,6 +1,0 @@
-<launch>
-	<node name="thrusters_serial_server" pkg="rosserial_python" type="serial_node.py" respawn="true">
-		<param name="port" value="/dev/auv_propulsion"/>
-		<param name="baud" value="57600"/>
-	</node>
-</launch>
