+++ catkin_ws/src/state_estimation/launch/state_estimation.launch
@@ -1,27 +1,12 @@
 <launch>
-	<include file="$(find imu)/launch/imu.launch"/>
-	<include file="$(find dvl)/launch/dvl.launch"/>
-	<include file="$(find state_estimation)/launch/transform_broadcaster.launch"/>
-	<include file="$(find state_estimation)/launch/state_aggregator.launch"/>
-
-
-	<!-- <node name="bag_record_bringup" pkg="rosbag" type="record" output="screen" args="-o $(find state_estimation)/rosbag/  
-		/pose
-		/pose_world
-		/state_x
-		/state_y
-		/state_z
-		/state_theta_x
-		/state_theta_y
-		/state_theta_z
-		/angular_velocity
-
-		/dead_reckon_report
-		/sbg/imu_data
-		/sbg/ekf_quat
-		/depth
-
-		/reset_state_full
-		/reset_state_planar "/> -->
-
+    <arg name="sim" default="false" />
+    <group unless="$(arg sim)">
+		<include file="$(find imu)/launch/imu.launch"/>
+		<include file="$(find dvl)/launch/dvl.launch"/>
+    </group>
+	<node name="transform_broadcaster" pkg="state_estimation" type="transform_broadcaster.py" respawn="false" output="screen"/>
+	<node name="state_aggregator" pkg="state_estimation" type="state_aggregator.py" respawn="true" output="screen"/>
+	<param name="theta_x_offset" type="double" value="0.0"/>
+	<param name="theta_y_offset" type="double" value="0.0"/>
+	<param name="theta_z_offset" type="double" value="0.0" />
 </launch>
