--- catkin_ws/src/state_estimation/launch/transform_broadcaster.launch
@@ -1,3 +1,0 @@
-<launch>
-	<node name="transform_broadcaster" pkg="state_estimation" type="transform_broadcaster.py" respawn="false" output="screen"/>
-</launch>
