+++ catkin_ws/src/vision/launch/gui_image_view.launch
@@ -1,3 +1,3 @@
 <launch>
-    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui" respawn="false" output="log" args = "--perspective-file $(find vision)/config/vision_gui.perspective"/>
+    <node name="rqt_gui" pkg="rqt_gui" type="rqt_gui" respawn="false" output="log" args = "--perspective-file $(find vision)/src/config/vision_gui.perspective"/>
 </launch>
