--- catkin_ws/src/vision/launch/sim_object_detection.launch
@@ -1,5 +1,0 @@
-<launch>
-    <node name="object_map" pkg="vision" type="object_map.py" respawn="true"  output="screen"/>
-    <node name="object_detection" pkg="vision" type="object_detection.py" respawn="true"  output="screen"/>
-    <!-- <node name="debug_thresholding" pkg="vision" type="debug_thresholding.py" respawn="true"  output="screen"/> -->
-</launch>
