+++ catkin_ws/src/vision/launch/vision.launch
@@ -1,0 +1,10 @@
+<launch>
+    <arg name="sim" default="false" />
+    <group unless="$(arg sim)">
+        <include file="$(find vision)/launch/stream-down-cam.launch" />
+        <!-- <include file="$(find vision)/launch/stream-front-cam.launch" /> -->
+    </group>
+    <node name="object_detection" pkg="vision" type="object_detection.py" respawn="true"  output="screen"/>
+    <node name="mapping" pkg="vision" type="object_map.py" respawn="true"  output="screen"/>
+    <!-- <node name="debug_thresholding" pkg="vision" type="debug_thresholding.py" respawn="true"  output="screen"/> -->
+</launch>
